# contracts/openapi/admin.yaml
openapi: 3.1.0
info:
  title: RingSays Admin API
  version: 0.2.0
  description: |
    API behind the enterprise portal and the RingSays back office.

    Authentication: OpenID Connect. The identity provider proves who the person is (`iss`, `sub`, verified
    email). RingSays decides what they may do: tenant memberships and roles are stored by RingSays, so a
    role change or a disabled user takes effect on the next request, not at token expiry.

    Tenant roles: TENANT_ADMIN, INTEGRATION_ADMIN, SUPERVISOR, AGENT, COMPLIANCE.
    RingSays staff roles: RS_REVIEWER (decides reviews), RS_ADMIN (creates and suspends tenants).

    Containment stays available while an organisation is suspended: revoking credentials, disabling
    webhook endpoints, stopping a calling number and disabling people.

    Back office paths (`/backoffice/...`, `/review/...`) are served only by the internal back office
    deployment, which holds the back office database role. The tenant facing deployment answers 404.

    Every write is recorded in the tenant's hash chained audit log with actor and, where required, reason.
    Phone numbers are never returned in full by this API.
servers:
  - url: https://api.ringsays.example/admin/v1
security:
  - oidc: []
...
paths:
...
  /backoffice/tenants:
    get:
      tags: [Back office]
      operationId: backofficeListTenants
      summary: All tenants in this region
      parameters:
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/TenantStatus" }
      responses:
        "200":
          description: Tenants.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Tenant" }
        "403": { $ref: "./common.yaml#/components/responses/Forbidden" }
        "404": { $ref: "./common.yaml#/components/responses/NotFound" }
    post:
      tags: [Back office]
      operationId: backofficeCreateTenant
      summary: Onboard an organisation and invite its first administrator (RS_ADMIN)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TenantCreate" }
      responses:
        "201":
          description: Created with status PENDING.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tenant" }
        "400": { $ref: "./common.yaml#/components/responses/BadRequest" }
        "403": { $ref: "./common.yaml#/components/responses/Forbidden" }
        "422": { $ref: "./common.yaml#/components/responses/UnprocessableEntity" }
  /backoffice/tenants/{tenant_id}/suspend:
    parameters:
      - $ref: "#/components/parameters/TenantId"
    post:
      tags: [Back office]
      operationId: backofficeSuspendTenant
      summary: Suspend; API access ends within 30 seconds (RS_ADMIN)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Reason" }
      responses:
        "200":
          description: Suspended.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tenant" }
        "403": { $ref: "./common.yaml#/components/responses/Forbidden" }
        "404": { $ref: "./common.yaml#/components/responses/NotFound" }
        "409": { $ref: "./common.yaml#/components/responses/Conflict" }
  /backoffice/tenants/{tenant_id}/reinstate:
    parameters:
      - $ref: "#/components/parameters/TenantId"
    post:
      tags: [Back office]
      operationId: backofficeReinstateTenant
      summary: Lift a suspension (RS_ADMIN)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Reason" }
      responses:
        "200":
          description: Reinstated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tenant" }
        "403": { $ref: "./common.yaml#/components/responses/Forbidden" }
        "404": { $ref: "./common.yaml#/components/responses/NotFound" }
        "409": { $ref: "./common.yaml#/components/responses/Conflict" }
  /review/queue:
    get:
      tags: [Back office]
      operationId: reviewQueue
      summary: Items waiting for a RingSays decision, oldest first
      responses:
        "200":
          description: Queue.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReviewQueue" }
        "403": { $ref: "./common.yaml#/components/responses/Forbidden" }
        "404": { $ref: "./common.yaml#/components/responses/NotFound" }
  /review/verifications/{request_id}:
    parameters:
      - name: request_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Back office]
      operationId: reviewGetVerification
      summary: Tenant profile and evidence for a verification request
      responses:
        "200":
          description: Request.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VerificationReview" }
        "403": { $ref: "./common.yaml#/components/responses/Forbidden" }
        "404": { $ref: "./common.yaml#/components/responses/NotFound" }
    post:
      tags: [Back office]
      operationId: reviewDecideVerification
      summary: Approve (tenant becomes VERIFIED) or reject with reason
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Decision" }
      responses:
        "200":
          description: Decision recorded.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VerificationRequest" }
        "403": { $ref: "./common.yaml#/components/responses/Forbidden" }
        "404": { $ref: "./common.yaml#/components/responses/NotFound" }
        "409": { $ref: "./common.yaml#/components/responses/Conflict" }
  /review/documents/{document_id}/file:
    parameters:
      - $ref: "#/components/parameters/DocumentId"
    get:
      tags: [Back office]
      operationId: reviewDownloadDocument
      summary: Download evidence file
      responses:
        "200":
          description: File, as uploaded. Download is recorded in the organisation's audit log.
          content:
            application/pdf:
              schema: { type: string, contentMediaType: application/pdf }
            image/png:
              schema: { type: string, contentMediaType: image/png }
            image/jpeg:
              schema: { type: string, contentMediaType: image/jpeg }
        "403": { $ref: "./common.yaml#/components/responses/Forbidden" }
        "404": { $ref: "./common.yaml#/components/responses/NotFound" }
  /review/purpose-codes/{tenant_id}/{code}:
    parameters:
      - $ref: "#/components/parameters/TenantId"
      - $ref: "#/components/parameters/Code"
    post:
      tags: [Back office]
      operationId: reviewPurposeCode
      summary: Approve or reject a proposed purpose code
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Decision" }
      responses:
        "200":
          description: Decision recorded.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminPurposeCode" }
        "403": { $ref: "./common.yaml#/components/responses/Forbidden" }
        "404": { $ref: "./common.yaml#/components/responses/NotFound" }
        "409": { $ref: "./common.yaml#/components/responses/Conflict" }
  /review/calling-numbers/{number_id}:
    parameters:
      - $ref: "#/components/parameters/NumberId"
    post:
      tags: [Back office]
      operationId: reviewCallingNumber
      summary: Approve or reject a calling number after checking operator registration
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Decision" }
      responses:
        "200":
          description: Decision recorded.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CallingNumber" }
        "403": { $ref: "./common.yaml#/components/responses/Forbidden" }
        "404": { $ref: "./common.yaml#/components/responses/NotFound" }
        "409": { $ref: "./common.yaml#/components/responses/Conflict" }
components:
...
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://auth.ringsays.example/.well-known/openid-configuration
  parameters:
    TenantId:
      name: tenant_id
      in: path
      required: true
      schema: { type: string, format: uuid }
    NumberId:
      name: number_id
      in: path
      required: true
      schema: { type: string, format: uuid }
    DocumentId:
      name: document_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Code:
      name: code
      in: path
      required: true
      schema: { type: string }
  schemas:
...
    TenantStatus:
      type: string
      enum: [PENDING, VERIFIED, SUSPENDED]
    Sector:
      type: string
      enum: [BANK, INSURANCE, FINANCE, GOVERNMENT]
...
    Reason:
      type: object
      additionalProperties: false
      required: [reason]
      properties:
        reason: { type: string, minLength: 3, maxLength: 500 }
    Decision:
      type: object
      additionalProperties: false
      required: [decision, reason]
      properties:
        decision: { type: string, enum: [APPROVE, REJECT] }
        reason: { type: string, minLength: 3, maxLength: 500 }
...
    Tenant:
      type: object
      required: [tenant_id, legal_name, sector, residency_region, verification_status, created_at]
      properties:
        tenant_id: { type: string, format: uuid }
        legal_name: { $ref: "./common.yaml#/components/schemas/LocalisedText" }
        commercial_registration: { type: [string, "null"] }
        domain: { type: [string, "null"] }
        sector: { $ref: "#/components/schemas/Sector" }
        residency_region: { type: string, enum: [KSA, IN, AE, EU] }
        verification_status: { $ref: "#/components/schemas/TenantStatus" }
        suspended_reason: { type: [string, "null"] }
        created_at: { type: string, format: date-time }
...
    TenantCreate:
      type: object
      additionalProperties: false
      required: [legal_name, sector, admin_email]
      properties:
        legal_name: { $ref: "./common.yaml#/components/schemas/LocalisedText" }
        sector: { $ref: "#/components/schemas/Sector" }
        domain: { type: string, maxLength: 253 }
        admin_email: { type: string, format: email, maxLength: 254 }
        admin_name: { type: string, maxLength: 120 }
...
    CallingNumber:
      type: object
      required: [number_id, phone_masked, status, cst_entity_name_registered, created_at]
      properties:
        number_id: { type: string, format: uuid }
        phone_masked: { type: string, description: "Business number with middle digits hidden, for example +9661•••••000." }
        department_id: { type: [string, "null"], format: uuid }
        status:
          type: string
          enum: [PENDING_VERIFICATION, VERIFIED, REVOKED]
        cst_entity_name_registered:
          type: boolean
          description: Whether number is registered in operator caller name programme.
        review_reason: { type: [string, "null"] }
        created_at: { type: string, format: date-time }
...
    DocumentKind:
      type: string
      enum: [COMMERCIAL_REGISTRATION, REGULATOR_LICENCE, AUTHORISATION_LETTER, DOMAIN_PROOF, OTHER]
    VerificationDocument:
      type: object
      required: [document_id, kind, file_name, content_type, size_bytes, sha256, uploaded_at]
      properties:
        document_id: { type: string, format: uuid }
        kind: { $ref: "#/components/schemas/DocumentKind" }
        reference: { type: [string, "null"] }
        file_name: { type: string }
        content_type: { type: string, enum: [application/pdf, image/png, image/jpeg] }
        size_bytes: { type: integer }
        sha256: { type: string }
        uploaded_at: { type: string, format: date-time }
    VerificationRequest:
      type: object
      required: [request_id, status, submitted_at]
      properties:
        request_id: { type: string, format: uuid }
        status: { type: string, enum: [SUBMITTED, APPROVED, REJECTED] }
        submitted_at: { type: string, format: date-time }
        decided_at: { type: [string, "null"], format: date-time }
        decision_reason: { type: [string, "null"] }
...
    VerificationReview:
      type: object
      required: [request, tenant, documents]
      properties:
        request: { $ref: "#/components/schemas/VerificationRequest" }
        tenant: { $ref: "#/components/schemas/Tenant" }
        documents:
          type: array
          items: { $ref: "#/components/schemas/VerificationDocument" }
...
    AdminPurposeCode:
      allOf:
        - $ref: "./enterprise.yaml#/components/schemas/PurposeCode"
        - type: object
          properties:
            proposed_at: { type: [string, "null"], format: date-time }
            reviewed_at: { type: [string, "null"], format: date-time }
            review_reason: { type: [string, "null"] }
...
    ReviewQueue:
      type: object
      required: [verifications, purpose_codes, calling_numbers]
      properties:
        verifications:
          type: array
          items:
            type: object
            required: [request_id, tenant_id, legal_name, sector, submitted_at]
            properties:
              request_id: { type: string, format: uuid }
              tenant_id: { type: string, format: uuid }
              legal_name: { $ref: "./common.yaml#/components/schemas/LocalisedText" }
              sector: { $ref: "#/components/schemas/Sector" }
              submitted_at: { type: string, format: date-time }
        purpose_codes:
          type: array
          items:
            allOf:
              - $ref: "#/components/schemas/AdminPurposeCode"
              - type: object
                required: [tenant_id, tenant_name]
                properties:
                  tenant_id: { type: string, format: uuid }
                  tenant_name: { $ref: "./common.yaml#/components/schemas/LocalisedText" }
        calling_numbers:
          type: array
          items:
            allOf:
              - $ref: "#/components/schemas/CallingNumber"
              - type: object
                required: [tenant_id, tenant_name, phone]
                properties:
                  tenant_id: { type: string, format: uuid }
                  tenant_name: { $ref: "./common.yaml#/components/schemas/LocalisedText" }
                  phone:
                    type: string
                    description: Full business number, shown to reviewers only, to check operator registration.